+++ Source/ACE.Entity/AceObject.cs
@@ -2228,8 +2228,8 @@ namespace ACE.Entity
         [JsonProperty("bookProperties")]
         public Dictionary<uint, AceObjectPropertiesBook> BookProperties { get; set; } = new Dictionary<uint, AceObjectPropertiesBook>();
 
-        [JsonProperty("generators")]
-        public List<AceObjectGeneratorLink> GeneratorLinks { get; set; } = new List<AceObjectGeneratorLink>();
+        //[JsonProperty("generators")]
+        //public List<AceObjectGeneratorLink> GeneratorLinks { get; set; } = new List<AceObjectGeneratorLink>();
 
         [JsonProperty("generatorProfiles")]
         public List<AceObjectGeneratorProfile> GeneratorProfiles { get; set; } = new List<AceObjectGeneratorProfile>();
@@ -2299,7 +2299,7 @@ namespace ACE.Entity
                 DataIdProperties = CloneList(DataIdProperties),
                 InstanceIdProperties = CloneList(InstanceIdProperties),
                 StringProperties = CloneList(StringProperties),
-                GeneratorLinks = CloneList(GeneratorLinks),
+                //GeneratorLinks = CloneList(GeneratorLinks),
                 GeneratorProfiles = CloneList(GeneratorProfiles),
                 CreateList = CloneList(CreateList),
                 AceObjectPropertiesAttributes = CloneDict(AceObjectPropertiesAttributes),
@@ -2336,12 +2336,12 @@ namespace ACE.Entity
             ret.DataIdProperties.ForEach(c => c.AceObjectId = guid);
             ret.InstanceIdProperties.ForEach(c => c.AceObjectId = guid);
             ret.StringProperties.ForEach(c => c.AceObjectId = guid);
-            ret.GeneratorLinks.ForEach(c => c.AceObjectId = guid);
-            ret.SpellIdProperties.ForEach(c => c.AceObjectId = guid);
-            ret.GeneratorLinks.ForEach(c => c.AceObjectId = guid);
+            //ret.GeneratorLinks.ForEach(c => c.AceObjectId = guid);
+            //ret.SpellIdProperties.ForEach(c => c.AceObjectId = guid);
+            //ret.GeneratorLinks.ForEach(c => c.AceObjectId = guid);
             ret.GeneratorProfiles.ForEach(c => c.AceObjectId = guid);
             ret.CreateList.ForEach(c => c.AceObjectId = guid);
-            ret.SpellsInSpellBars.ForEach(c => c.AceObjectId = guid);
+            //ret.SpellsInSpellBars.ForEach(c => c.AceObjectId = guid);
             // Cloning an object as new should not clone inventory I don't think intentionally left out. Og II
 
             // No need to change Dictionary guids per DDEVEC